# back/build/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Mindi_API
  description: |
    # Mindi의 API 명세서
servers:
  - url: 'http://localhost:5001/api/'
...
paths:
...
  /posts:
    post:
      tags:
        - posts
      summary: 새로운 게시글을 등록합니다.
      security:
        - BearerAuth: []
      description: |
        # 게시글 업로드 API
        새로운 게시글을 등록합니다.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
              properties:
                title:
                  type: string
                content:
                  type: string
            examples:
              value:
                title: 등록할 글 제목
                content: 등록할 글 내용
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: 요청이 성공했는지를 나타냅니다.
                  post:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: postId
                      title:
                        type: string
                        description: 글 제목
                      content:
                        type: string
                        description: 글 내용
                      author:
                        type: string
                        description: 작성자 userId
                      createdAt:
                        type: string
                        description: 작성일
              examples:
                value:
                  success: true
                  post:
                    _id: 62a0c9fe574953ef39d5c9e8
                    title: 게시글 제목
                    content: 게시글 내용
                    author: 62a0c1234jk4953ef39d5c9e9
                    createdAt: 2022-06-27T16:26:08.844Z
    get:
      tags:
        - posts
      summary: 게시글 목록을 가져옵니다.
      security:
        - BearerAuth: []
      description: |
        # 게시글 목록 API
        게시글 목록을 반환합니다.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          description: 페이지 넘버
        - name: limit
          in: query
          schema:
            type: integer
          description: 한 페이지에 표시할 게시글 수
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: 요청이 성공했는지를 나타냅니다.
                  posts:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: postId
                        title:
                          type: string
                          description: 게시글 제목
                        content:
                          type: string
                          description: 게시글 내용
                        author:
                          type: string
                          description: 게시글 작성자 userId
                        comments:
                          type: number
                          description: 게시글 댓글 개수
                        createdAt:
                          type: string
                          description: 게시글 작성일
              examples:
                value:
                  success: true
                  posts:
                    - _id: 62a0c9fe574953ef39d5c9e8
                      title: 게시글 제목1
                      content: 게시글 내용1
                      author: 62a0c1234jk4953ef39d5c9e9
                      createdAt: 2022-06-27T16:26:08.844Z
                    - _id: 62a0c9fe574953ef39d5c9e8
                      title: 게시글 제목2
                      content: 게시글 내용2
                      author: 62a0c1234jk4953ef39d5c9e9
                      createdAt: 2022-06-27T16:26:08.844Z
  '/posts/{postId}':
    get:
      tags:
        - posts
      summary: 게시글 정보를 가져옵니다.
      security:
        - BearerAuth: []
      description: |
        # 게시글 정보 열람 API
        게시글 1개의 정보를 반환합니다.
      parameters:
        - name: postId
          in: path
          description: postId
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: 요청이 성공했는지를 나타냅니다.
                  post:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: postId
                      title:
                        type: string
                        description: 글 제목
                      content:
                        type: string
                        description: 글 내용
                      author:
                        type: string
                        description: 작성자 userId
                      comments:
                        type: number
                        description: 댓글 개수
                      createdAt:
                        type: string
                        description: 작성일
              examples:
                value:
                  success: true
                  posts:
                    _id: 62a0c9fe574953ef39d5c9e8
                    title: 게시글 제목
                    content: 게시글 내용
                    author: 62a0c1234jk4953ef39d5c9e9
                    createdAt: 2022-06-27T16:26:08.844Z
    put:
      tags:
        - posts
      summary: 게시글 정보를 수정합니다.
      security:
        - BearerAuth: []
      description: |
        # 게시글 수정 API
        게시글을 수정합니다.
      parameters:
        - name: postId
          in: path
          description: postId
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
              properties:
                title:
                  type: string
                content:
                  type: string
            examples:
              value:
                title: 수정할 글 제목
                content: 수정할 글 내용
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: 요청이 성공했는지를 나타냅니다.
                  post:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: postId
                      title:
                        type: string
                        description: 글 제목
                      content:
                        type: string
                        description: 글 내용
                      author:
                        type: string
                        description: 작성자 userId
                      createdAt:
                        type: string
                        description: 작성일
              examples:
                value:
                  success: true
                  post:
                    _id: 62a0c9fe574953ef39d5c9e8
                    title: 수정된 게시글 제목
                    content: 수정된 게시글 내용
                    author: 62a0c1234jk4953ef39d5c9e9
                    createdAt: 2022-06-27T16:26:08.844Z
    delete:
      tags:
        - posts
      summary: 게시글을 삭제합니다.
      security:
        - BearerAuth: []
      description: |
        # 게시글 삭제 API
        게시글 1개를 삭제합니다.
      parameters:
        - name: postId
          in: path
          description: postId
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: 요청이 성공했는지를 나타냅니다.
                  message:
                    type: string
                    descipriton: 메시지
              examples:
                value:
                  success: true
                  message: 성공적으로 삭제되었습니다.
  /users/posts:
    get:
      tags:
        - posts
      summary: 게시글 목록을 가져옵니다.
      security:
        - BearerAuth: []
      description: |
        # 유저 게시글 목록 API
        로그인한 사용자의 게시글 목록을 반환합니다.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          description: 페이지 넘버
        - name: limit
          in: query
          schema:
            type: integer
          description: 한 페이지에 표시할 게시글 수
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: 요청이 성공했는지를 나타냅니다.
                  posts:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: postId
                        title:
                          type: string
                          description: 게시글 제목
                        content:
                          type: string
                          description: 게시글 내용
                        author:
                          type: string
                          description: 게시글 작성자 userId
                        comments:
                          type: number
                          description: 게시글 댓글 개수
                        createdAt:
                          type: string
                          description: 게시글 작성일
              examples:
                value:
                  success: true
                  posts:
                    - _id: 62a0c9fe574953ef39d5c9e8
                      title: 게시글 제목1
                      content: 게시글 내용1
                      author: 62a0c1234jk4953ef39d5c9e9
                      createdAt: 2022-06-27T16:26:08.844Z
                    - _id: 62a0c9fe574953ef39d5c9e8
                      title: 게시글 제목2
                      content: 게시글 내용2
                      author: 62a0c1234jk4953ef39d5c9e9
                      createdAt: 2022-06-27T16:26:08.844Z
  '/users/posts/{userId}':
    get:
      tags:
        - posts
      summary: 게시글 목록을 가져옵니다.
      security:
        - BearerAuth: []
      description: |
        # 유저 게시글 목록 API
        로그인한 사용자의 게시글 목록을 반환합니다.
      parameters:
        - name: userId
          in: path
          schema:
            type: string
          description: userId
        - name: page
          in: query
          schema:
            type: integer
          description: 페이지 넘버
        - name: limit
          in: query
          schema:
            type: integer
          description: 한 페이지에 표시할 게시글 수
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: 요청이 성공했는지를 나타냅니다.
                  posts:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: postId
                        title:
                          type: string
                          description: 게시글 제목
                        content:
                          type: string
                          description: 게시글 내용
                        author:
                          type: string
                          description: 게시글 작성자 userId
                        comments:
                          type: number
                          description: 게시글 댓글 개수
                        createdAt:
                          type: string
                          description: 게시글 작성일
              examples:
                value:
                  success: true
                  posts:
                    - _id: 62a0c9fe574953ef39d5c9e8
                      title: 게시글 제목1
                      content: 게시글 내용1
                      author: 62a0c1234jk4953ef39d5c9e9
                      createdAt: 2022-06-27T16:26:08.844Z
                    - _id: 62a0c9fe574953ef39d5c9e8
                      title: 게시글 제목2
                      content: 게시글 내용2
                      author: 62a0c1234jk4953ef39d5c9e9
                      createdAt: 2022-06-27T16:26:08.844Z
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer